+++ src/Behaviours/PersisterActions/HasRouteData.php
@@ -2,6 +2,11 @@
 
 namespace Somnambulist\ApiClient\Behaviours\PersisterActions;
 
+use function in_array;
+use function is_null;
+use function strtolower;
+use function trim;
+
 /**
  * Trait HasRouteData
  *
@@ -25,10 +30,27 @@ trait HasRouteData
      */
     protected $params;
 
-    public function route(string $route, array $params = []): self
+    /**
+     * HTTP method
+     *
+     * @var string|null
+     */
+    protected $method;
+
+    protected function validateHttpMethod(?string $method): ?string
+    {
+        if (is_null($method)) {
+            return null;
+        }
+
+        return in_array(strtolower(trim($method)), ['get', 'post', 'put', 'patch', 'delete', 'head']) ? $method : null;
+    }
+
+    public function route(string $route, array $params = [], string $method = null): self
     {
         $this->route  = $route;
         $this->params = $params;
+        $this->method = $this->validateHttpMethod($method);
 
         return $this;
     }
@@ -41,5 +63,10 @@ trait HasRouteData
     public function getRouteParams(): array
     {
         return $this->params;
+    }
+
+    public function getMethod(): ?string
+    {
+        return $this->method;
     }
 }